Refactor : test를 클래스별로 나누고, 예상 하지 못한 exception은 오류가 뜨도록 코드 수정
